discord-evaluator/requirements.txt
discord.py==2.6.4
docker_py==1.10.6
python-dotenv==1.2.1
